Separated logging into platform specific code.
For Linux output Chalk is used to colorize the output. To prevent those
character codes from being displayed in the windows terminal the core
log code has been separated so it can be handled differently on windows
